refactor: Controller 명칭 변경
