+++ src/test/java/com/givemecon/controller/ApiExceptionControllerTest.java
@@ -1,4 +1,4 @@
-package com.givemecon.controller.api;
+package com.givemecon.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.givemecon.config.auth.dto.TokenInfo;
@@ -32,7 +32,7 @@ import java.time.LocalDate;
 
 import static com.givemecon.config.enums.JwtAuthHeader.*;
 import static com.givemecon.config.enums.Authority.*;
-import static com.givemecon.controller.TokenHeaderUtils.getAccessTokenHeader;
+import static com.givemecon.util.TokenHeaderUtils.getAccessTokenHeader;
 import static com.givemecon.domain.member.MemberDto.*;
 import static com.givemecon.domain.voucher.VoucherDto.*;
 import static com.givemecon.domain.voucher.VoucherStatus.FOR_SALE;
