+++ src/test/java/com/givemecon/controller/VoucherControllerTest.java
@@ -1,4 +1,4 @@
-package com.givemecon.controller.api;
+package com.givemecon.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.givemecon.config.auth.dto.TokenInfo;
@@ -44,8 +44,8 @@ import java.util.List;
 
 import static com.givemecon.config.enums.JwtAuthHeader.*;
 import static com.givemecon.config.enums.Authority.*;
-import static com.givemecon.controller.ApiDocumentUtils.*;
-import static com.givemecon.controller.TokenHeaderUtils.getAccessTokenHeader;
+import static com.givemecon.util.ApiDocumentUtils.*;
+import static com.givemecon.util.TokenHeaderUtils.getAccessTokenHeader;
 import static com.givemecon.domain.member.MemberDto.*;
 import static com.givemecon.domain.voucher.VoucherDto.*;
 import static com.givemecon.domain.voucher.VoucherStatus.*;
@@ -65,7 +65,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 @Import(S3MockConfig.class)
 @Transactional
 @SpringBootTest
-class VoucherApiControllerTest {
+class VoucherControllerTest {
 
     @Autowired
     WebApplicationContext context;
