+++ src/test/java/com/givemecon/controller/jwt/TokenIssueApiControllerTest.java
@@ -27,8 +27,8 @@ import java.util.UUID;
 import static com.givemecon.config.enums.JwtAuthHeader.*;
 import static com.givemecon.config.enums.Authority.*;
 import static com.givemecon.config.enums.OAuth2ParameterName.*;
-import static com.givemecon.controller.ApiDocumentUtils.*;
-import static com.givemecon.controller.TokenHeaderUtils.*;
+import static com.givemecon.util.ApiDocumentUtils.*;
+import static com.givemecon.util.TokenHeaderUtils.*;
 import static com.givemecon.domain.member.MemberDto.*;
 import static org.assertj.core.api.Assertions.*;
 import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
